+++ sga/models.py
@@ -2,7 +2,7 @@ from django.contrib.auth.models import User
 from django.core.validators import MinValueValidator, MaxValueValidator
 from django.db import models
 
-from sga.backend.constants import student_submission_file_path, grader_submission_file_path
+from sga.backend.files import student_submission_file_path, grader_submission_file_path
 from sga.backend.validators import validate_file_extension
 
 
@@ -59,7 +59,6 @@ class Student(models.Model):
   
 class Course(TimeStampedModel):
     edx_id = models.CharField(max_length=128, unique=True)
-    name = models.CharField(max_length=128)
     administrators = models.ManyToManyField(User, related_name="administrator_courses")
     graders = models.ManyToManyField(User, through=Grader, related_name="grader_courses")
     students = models.ManyToManyField(User, through=Student, related_name="student_courses")
