+++ sga/templates/sga/view_assignment.html
@@ -16,7 +16,7 @@
 {% endblock %}
 
 {% block content %}
-    <h1>Course: {{ course.name }}</h1>
+    <h1>Course: {{ course.edx_id }}</h1>
     <h3>Assignment: {{ assignment.name }}</h3>
     <hr>
     <table id="student-list" class="table table-striped table-hover">
