+++ sga/templates/sga/view_grader.html
@@ -34,7 +34,7 @@
 {% endblock %}
 
 {% block content %}
-    <h1>Course: {{ course.name }}</h1>
+    <h1>Course: {{ course.edx_id }}</h1>
     <h3>Grader: {{ grader }}</h3>
     
     <hr>
