+++ sga/templates/sga/view_grader_list.html
@@ -25,7 +25,7 @@
 {% endblock %}
 
 {% block content %}
-    <h1>Course: {{ course.name }}</h1>
+    <h1>Course: {{ course.edx_id }}</h1>
     <h3>Grader List</h3>
     <hr>
     <h4>Filters</h4>
