+++ sga/templates/sga/view_student.html
@@ -28,7 +28,7 @@
 {% endblock %}
 
 {% block content %}
-    <h1>Course: {{ course.name }}</h1>
+    <h1>Course: {{ course.edx_id }}</h1>
     <h3>Student: {{ student }}</h3>
     
     <hr>
