+++ sga/templates/sga/view_student_list.html
@@ -17,7 +17,7 @@
 {% endblock %}
 
 {% block content %}
-    <h1>Course: {{ course.name }}</h1>
+    <h1>Course: {{ course.edx_id }}</h1>
     {% if grader_user %}
     <h3>Student List (Grader: {{ grader_user.get_full_name }})</h3>
     {% else %}
